+++ src/HAMqtt.cpp
@@ -20,12 +20,10 @@ void onMessageReceived(char* topic, uint8_t* payload, unsigned int length)
 HAMqtt::HAMqtt(Client& netClient, HADevice& device) :
     _netClient(netClient),
     _device(device),
-    _hasDevice(true),
+    _connectedCallback(nullptr),
     _initialized(false),
     _discoveryPrefix(DefaultDiscoveryPrefix),
     _mqtt(new PubSubClient(netClient)),
-    _serverIp(new IPAddress()),
-    _serverPort(0),
     _username(nullptr),
     _password(nullptr),
     _lastConnectionAttemptAt(0),
@@ -67,13 +65,11 @@ bool HAMqtt::begin(
         return false;
     }
 
-    *_serverIp = serverIp;
-    _serverPort = serverPort;
     _username = username;
     _password = password;
     _initialized = true;
 
-    _mqtt->setServer(*_serverIp, _serverPort);
+    _mqtt->setServer(serverIp, serverPort);
     _mqtt->setCallback(onMessageReceived);
 
     return true;
@@ -85,16 +81,78 @@ bool HAMqtt::begin(
     const char* password
 )
 {
-    return begin(serverIp, 1883, username, password);
+    return begin(serverIp, HAMQTT_DEFAULT_PORT, username, password);
+}
+
+bool HAMqtt::begin(
+    const char* hostname,
+    const uint16_t& serverPort,
+    const char* username,
+    const char* password
+)
+{
+#if defined(ARDUINOHA_DEBUG)
+    Serial.println(F("Initializing ArduinoHA"));
+    Serial.print(F("Server address: "));
+    Serial.print(hostname);
+    Serial.print(F(":"));
+    Serial.print(serverPort);
+    Serial.println();
+#endif
+
+    if (_device.getUniqueId() == nullptr) {
+#if defined(ARDUINOHA_DEBUG)
+        Serial.println(F("Failed to initialize ArduinoHA. Missing device's unique ID."));
+#endif
+
+        return false;
+    }
+
+    if (_initialized) {
+#if defined(ARDUINOHA_DEBUG)
+    Serial.println(F("ArduinoHA is already initialized"));
+#endif
+
+        return false;
+    }
+
+    _username = username;
+    _password = password;
+    _initialized = true;
+
+    _mqtt->setServer(hostname, serverPort);
+    _mqtt->setCallback(onMessageReceived);
+
+    return true;
+}
+
+bool HAMqtt::begin(
+    const char* hostname,
+    const char* username,
+    const char* password
+)
+{
+    return begin(hostname, HAMQTT_DEFAULT_PORT, username, password);
+}
+
+bool HAMqtt::disconnect()
+{
+    if (!_initialized) {
+        return false;
+    }
+
+    // publish LWT
+
+    _initialized = false;
+    _lastConnectionAttemptAt = 0;
+    _mqtt->disconnect();
+
+    return true;
 }
 
 void HAMqtt::loop()
 {
-    if (!_initialized) {
-        return;
-    }
-
-    if (!_mqtt->loop()) {
+    if (_initialized && !_mqtt->loop()) {
         connectToServer();
     }
 }
@@ -224,7 +282,11 @@ void HAMqtt::connectToServer()
         Serial.println(F("Connected to the broker"));
 #endif
 
-        onConnected();
+        notifyDevicesAboutConnection();
+
+        if (_connectedCallback) {
+            _connectedCallback();
+        }
     } else {
 #if defined(ARDUINOHA_DEBUG)
         Serial.println(F("Failed to connect to the broker"));
@@ -232,7 +294,7 @@ void HAMqtt::connectToServer()
     }
 }
 
-void HAMqtt::onConnected()
+void HAMqtt::notifyDevicesAboutConnection()
 {
     for (uint8_t i = 0; i < _devicesTypesNb; i++) {
         _devicesTypes[i]->onMqttConnected();