+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * Added support for setting retain flag in HASwitch
 * Added support for text (const char*) payload in HASensor
 * Added support for fans (HAFan)
+* Added support for connecting to the MQTT broker using hostname
+* Added `onConnected()` method in the HAMqtt
 
 **Updates:**
 * Optimized codebase and logic in all devices types